requirements.txt
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
transformers
accelerate
torch
pypdf
gradio
bitsandbytes